+++ kernel/include/disk/disk.h
@@ -62,7 +62,7 @@ typedef enum {
 	ATA_PIO_SR_BSY			= 0x80
 } ATA_PIO_STATUS_REG;
 
-uint8_t read_sectors(uint8_t, uint8_t, uint32_t);
-uint8_t write_sectors(uint8_t, uint8_t, uint32_t);
+uint8_t read_sectors(uint32_t, uint32_t, uint32_t);
+uint8_t write_sectors(uint32_t, uint32_t, uint32_t);
 
 #endif /* !DISK_H */
